Add a basic test framework using HUnit.  Split the parsing monad into its own file.
Make the parsing monad a real instance of Monad.  It looks really
basic right now, but it should be easier to add support for line
number tracking this way.

Build the tests using:

  cabal configure -fbuildTests

The analysis bits are now a library used by the tester.  Any real
analyses can later just be executables using this library.  This lets
the library code be shared in a useful way (new to Cabal 1.8) so that
every single library module doesn't need to be rebuilt for the tester.
This saves a lot of time.