Change things up to process all metadata before values.
The few values refered to from metadata are "backwards".  The function
definition metadata, for example, refers to a Function value.  This is
not entirely useful.  Now, assemble a separate map of metadata that is
intended to be attached to a value but is not normally specified that way:

 * Global variables
 * Function definitions

These pieces of metadata will be associated with their actual values
once they are constructed in the next pass.

This actually compiles and should be reasonably complete - some
translation of constant values needs to be performed still for one
piece of metadata (marked with a FIXME).  This will be done after I
write that code for the generic Value translator (next step).