Undo the ABI-lowering inversion in the function type generator.
On further reflection, undoing the ABI lowering is difficult
and probably not a good idea.  LLVM's SSA form is designed so
that no memory values ever appear as SSA values.  This would
have to be violated to make the ABI-unlowering transformations.

Dealing with those at the analysis level (if even required) is
not very difficult with appropriate abstractions.