Add a little test driver that dumps argument types (from debug info).
This shows that the debug info attachment process works, including for
the very complex case of function arguments.  This is also the
information that I was most interested in.

The driver isn't complete and just dumps a raw representation instead
of chasing through all of the possible type description records.