Strictify basically everything.  After tying the knot on the Module, perform the equivalent of a deepseq.
This code doesn't use deepseq all the way down because the Module has
cycles, but it performs a safe equivalent.  It might miss a few minor
things and doesn't touch Types, but none of that stuff takes much
space so it doesn't matter.

Judging by heap profiles, the knot tying phase is very easy to spot
and the intermediate data is all dropped as exepected.