A lot more translated, various errors fixed.
Now the problem is using addresses as unique IDs.  These are not
compatible with Int, which is a really handy property to have for
various other purposes.  I'll have to switch back to Int somehow -
possibly by just mapping addresses to unique IDs with an explicit Map.
Keeping the splittable streams is also possible, though more
challenging.