Note extra files that are required in cabal.  Give names to instructions (LLVM doesn't store them explicitly).
The names take up some space - it may be worth leaving them anonymous in the future
and just assigning names when printing things out.  That said, it is pretty convenient
to be able to just print a value standalone, so I am inclined to keep them.  It may
be worth converting them to SmallStrings though