Add a fix for the construction of NamedTypes.
This was a big problem - before all named types were created as Void
since the tag was never set.  This sets the tag.  There was also a bug
in the translation of named types since it was never tested due to the
other bug.  In some patterns of struct reference, the named type would
never get initially translated, leading to a knot tying failure.