Efficiency and correctness work on the ICFG.
Adding a (hopefully temporary) fork of the PatriciaTree Graph
implementation from FGL.  It is modified to not retain a reference to
the fully evaluated edge list, which saves a ton of memory (the list
can be streamed).

Also improve the precision of the trivial function pointer analysis in
easy cases.

gnutls takes almost no memory above the Module representation now.