Start stubbing out the interface for IR modification.
This uses data-lens for field access.  Despite the lack of examples
the library is very nice and makes updating fields in the State monad
very easy.  Unfortunately, the Template Haskell helper package can't
be used since it dynamically loads C++ code.  The compiler seems to just
die on that, unfortunately.