Change the expected results of some tests.
The results of the analysis were correct because arguments do not have
any storage - the compiler will drop them down to stack slots
manually (the %1 refs in the results).  It takes an extra layer on top
of the points-to analysis to map things to argument names and that is
fine.