Do not return functions from the dataflow analysis
Previously, the dataflow analysis was meant to be partially applied
and the result used to look up values (by feeding it an Instruction).
This is unintuitive and causes some problems when combined with the
CallGraphSCCTraversal.  In that configuration, the dataflow analysis
is meant to provide a summary value for each function, with the
CallGraphSCCTraversal applying it in the best order possible.

To determine when fixed points are reached in SCCs, the traversal
compares the summary values for equality.  When the dataflow analysis
returned functions, this comparison was impossible (there is no way to
compare functions for equality in Haskell).

This change just returns a data structure (HashMap) mapping
Instructions to their dataflow facts.  It required a bit of mucking
around in the internals of the dataflow solver to work in terms of
Instructions instead of CFG Nodes, but the changes seem to work (the
escape analysis tests still pass).  There were some other related
cleanups in the dataflow solver to make that all nicer.