Make the CallGraphSCCTraversal monadic
This allows the underlying analysis to use something like the RWS (or
just reader and writer monads) to log data and pass global
information.

If an analysis doesn't need those things, just use runIdentity.

The dataflow analysis framework still needs to be modified.  That is
the real target, so that warnings can be logged &c.