Cluster nodes by basic block in the CDG
Instructions in the same basic block should all have the same control
dependencies.  The clustering makes the graph layout *much* faster by
removing degrees of freedom.

Note: The current crash in the CDG construction code happens in
functions that do not return via a ret instruction.  Examples: longjmp
and exit.  The problem is that the postdominator tree isn't a tree
because the wrong exit instruction is chosen.

To fix this, do a real analysis to find the real exit instruction.
For now, assume that there is a unique exit.... but that isn't
necessarily the case.