Remove the old instruction-granularity dataflow analysis, replace with block level.
The instruction-level one was not correct anymore w.r.t. phi
handling (among other things).  The block-level analysis is correct
and this change modifies it to store results in the map for each
instruction.

This greatly simplifies the dataflowResult function since it doesn't
need to apply the transfer function anymore.  In turn, that removes a
lot of code duplication and everything is more robust now.