Add more tests for the escape analysis (with a fix)
This includes a fix to the analysis for one of the tests.  The problem
was that the DFS always reports that a node is reachable from
itself (a problem for the instructionEscapes function).  The quick
hack is to remove the instruction from the set of reachable nodes
before doing the standard check.  The problem is that a function
result could be passed back to it indirectly (via a phi node and a
loop), though which it could escape.

The right fix is to first do a check on the graph to see if the input
instruction is part of a cycle.  If it is, don't remove it.