Abstract Function inputs in analyses, rewrite the CallGraphSCCTraversal
The traversal is now more generic - there are easy adapters to make
pure or monadic analyses.  There is also a variant for *composable*
analyses.  These are analyses that are run in parallel over the call
graph.  That is, all analyses are run over a given SCC before moving
on to the next SCC.

This is much more efficient than running them separately over the CFG;
in particular, it makes it much easier to share intermediate
values (like a compute CFG or CDG).