Add a helper 'followAccessPath'
Given an AbstractAccessPath and a Value, it will do its best to follow
the path through the value, returning the value described by the path
(assuming it was rooted at the input value).  The function will only
return a value if the input is a constant and the path only touches
constants/globals.