Rewrite the Class Hierarchy Analysis to deal with virtual inheritance
Virtual inheritance introduces complex initialization and construction
code that is not amenable to the old "inspect constructors" approach
to analysis.  The new code uses C++ run-time type information embedded
in Modules to reconstruct the class hierarchy instead.

This is robust to virtual inheritance and also more normal cases.  It
requires a bit more care in dealing with names and further testing is
required.  Basic tests of namespace handling and virtual inheritance
pass.