Remove some partial code and remove unused cases in the escape analysis
There was no need for a separate FieldSink constructor (eliminated).
Some partial cases were eliminated (should be all in practice).

This change also fixes a bug where one escaping argument in a function
would cause all of its fptr escapes to be overwritten with escapes.

The case where an argument escapes into a field of another argument
presents an interesting challenge since there is no obvious store
instruction to flag as the witness.  This code now searches for all of
the store instructions along the path from the argument to the sink to
have a full witness trace.  Currently, only the first store is reported.