Revive the andersens analysis
This time the implementation is based on an actual set constraints
solver (which is external and can be optimized separately).  Several
tests pass and the basic operation seems to be in place.  It needs
handling of phi nodes and function calls, but other than that seems
to be well on its way.

Field-sensitivity will come later; field sensitivity may be limited in
some ways.  Additionally, the analysis will have special cases to
handle C++ virtual function calls.