Add more cases for the points-to query
If the value given for the points-to query is represented by a special
variable in the constraint graph (e.g., ret nodes, phi nodes, etc),
then the query needs to use the same type of special variable.

The old code only had special cases for call return values and loads;
this commit adds the rest.  The GEP rule might not be correct (it
should probably be a FieldLoc node, as noted in the code).