Fixes to the callGraphSCCTraversal
1) Remove the special case for handling single functions.  This
special case was avoiding computing a fixed point for single function
SCCs, but this is not correct if the function has a self-loop.  For
now, always compute a fixed point even if a single function SCC does
not call itself.  This can be optimized later, but for now it is more
correct.

2) Fill in some missing cases for *pure* analysis functions.  I was
only using the monadic forms so I had skipped these before.