Track the baseType of an AccessPath separately from its baseValue
This is a minor tweak, but it allows us to more easily work with paths
that begin with a value that is bitcast to another type.  Before,
paths would look like they started at the type pre-bitcast.  Now the
recorded base value will be the same but the recorded base type is the
type cast *to*.

This may seem odd, but if the value is explicitly cast *to* something
else, that is much more likely to be its true type.

It may be prudent to split this into a different function so that both
behaviors are available if required.