Fixes to the backwards dataflow analysis and the fixpoint computation
Backwards dataflow was not computing per-instruction results correctly
in dataflowResultAt, which required storing the direction in the
dataflow result so appropriate corrections could be made.  Namely, the
input fact had to be computed differently and instructions in the
basic block must be traversed in /reverse/ order.

The fixpoint problem arose when the first transfer function produced
an output fact that was the same as its initial value.  The loop saw
no change and so added no successors.  Now we track which instructions
we have seen - if we see an instruction for the first time, we always
add its successors regardless of whether or not it changed.

The original hoopl code handled this implicitly in the graph rewriting
step (a graph was rewritten to itself, I think, and so the necessary
nodes were added to the worklist?)