Bugfix for the BLockReturnValue analysis
Before, querying for the returned values of a block would return partial
results in some cases.  The root cause is that return values aren't
pushed back past certain dominance relations.

Not pushing is correct, but it caused misleading values to be returned.
Now, if a value is not pushed back because of dominance, that path is
"poisoned" and queries where information would be missing now return
Nothing.