Add a uses-of analysis
This information is available in LLVM and is occasionally useful.
Storing it in Values directly would be difficult with this IR,
though, so we store it on the side and it can be computed as-needed.

It is computed for an entire module at a time, so cross-function
information is available.